// src/helpers/config.d.ts
export declare const POOL_INDEX_CONTRACT_ID = "G2j_YAD1GQcdtXZEwUIE7VDs8Y0UuWx85inKI-kXajY";
export declare const TAGS: {
    keys: {
        alexPoolId: string;
        ansTitle: string;
        ansDescription: string;
        ansTopic: string;
        ansType: string;
        ansImplements: string;
        appType: string;
        artifactName: string;
        artifactType: string;
        associationId: string;
        associationSequence: string;
        childAssets: string;
        bookmarkIds: string;
        bookmarkSearch: string;
        dateCreated: string;
        keywords: string;
        initialOwner: string;
        poolId: string;
        profileImage: string;
        uploaderTxId: string;
        contractSrc: string;
        mediaIds: string;
        timestamp: string;
        collectionName: string;
        collectionDescription: string;
    };
    values: {
        defaultArtifactType: string;
        imageArtifactType: string;
        messagingArtifactType: string;
        nostrEventArtifactType: string;
        redditThreadArtifactType: string;
        webpageArtifactType: string;
        poolVersions: {
            '1.2': string;
            '1.4': string;
            '1.5': string;
        };
        searchIndex: string;
        collectionAppType: string;
        ansVersion: string;
        ansType: string;
    };
};
export declare const STORAGE: {
    none: string;
};
export declare const PAGINATOR = 100;
export declare const CURSORS: {
    p1: string;
    end: string;
};
export declare const MEDIA_TYPES: {
    mp4: string;
    jpg: string;
    jpeg: string;
    png: string;
};
export declare const SEARCH: {
    cursorPrefix: string;
    idTerm: string;
    ownerTerm: string;
};
export declare const FALLBACK_IMAGE = "8HqSqy_nNRSTPv-q-j7_iHGTp6lEA5K77TP4BPuXGyA";
export declare const OPERATOR_LINK = "https://alex-operator-guide.arweave.dev/";