// src/helpers/types.d.ts
export declare enum ArtifactEnum {
    Messaging = "Alex-Messaging",
    Webpage = "Alex-Webpage",
    Reddit = "Alex-Reddit-Thread",
    Nostr = "Alex-Nostr-Event"
}
export declare enum CursorEnum {
    GQL = "gql",
    Search = "search"
}
export type GQLResponseType = {
    cursor: string | null;
    node: {
        id: string;
        tags: KeyValueType[];
        data: {
            size: string;
            type: string;
        };
    };
};
export interface ArtifactDetailType {
    artifactId: NStringType;
    artifactName: NStringType;
    artifactType: ArtifactEnum.Messaging | ArtifactEnum.Webpage | ArtifactEnum.Reddit | ArtifactEnum.Nostr;
    associationId: NStringType;
    associationSequence: NStringType;
    profileImagePath: NStringType;
    owner: NStringType;
    ansTitle: NStringType;
    minted: NStringType;
    keywords: NStringType;
    poolName: NStringType;
    mediaIds: NStringType;
    childAssets: NStringType;
    poolId: NStringType;
    dataUrl: NStringType;
    dataSize: NStringType;
    rawData: NStringType;
}
export interface AssociationDetailType {
    artifacts: ArtifactDetailType[];
    length: number;
}
export type ArtifactArgsType = {
    ids: string[] | null;
    owner: string | null;
    uploader: string | null;
    cursor: string | null;
    reduxCursor: string | null;
};
export type ArcGQLResponseType = {
    data: GQLResponseType[];
    nextCursor: string | null;
};
export type ArtifactResponseType = {
    nextCursor: string | null;
    previousCursor: string | null;
    contracts: GQLResponseType[];
};
export interface PoolType {
    id: string;
    state: PoolStateType;
}
export interface PoolStateType {
    title: string;
    image: string;
    briefDescription: string;
    description: string;
    link: string;
    owner: string;
    ownerInfo: string;
    timestamp: string;
    contributors: {
        [key: string]: string;
    };
    tokens: {
        [key: string]: string;
    };
    totalContributions: string;
    totalSupply: string;
    balance: string;
}
// ...
export type NStringType = string | null;
export type NStringListType = string[] | null;
export type URLViewType = {
    index: number;
    label: string;
    icon: string;
    disabled: boolean;
    url: any;
    view: any;
};
export interface IURLView {
    account: URLViewType[];
    library: URLViewType[];
}
export type ValidationType = {
    status: boolean;
    message: string | null;
};
export type KeyValueType = {
    [key: string]: string | any;
};
// ...
export type NotificationResponseType = {
    status: number | null;
    message: string | null;
};
export type TableIdType = {
    value: string;
    type: 'poolId' | 'ownerId';
};
export type TagFilterType = {
    name: string;
    values: string[];
};
// ...
export type SequenceType = {
    start: number;
    end: number;
};